MVC changes to show data on jsp page

+++ src/main/java/com/docsolr/entity/SiteFolders.java
@@ -9,7 +9,7 @@ import javax.persistence.*;
 public class SiteFolders extends BaseEntity {
 	
 	private String uniqueId;
-	private int siteLibraryId;
+	private long siteLibraryId;
 
 	private String name;
 	private Date timeLastModified;
@@ -19,11 +19,11 @@ public class SiteFolders extends BaseEntity {
 	private String siteName;
 	private String hostURL;
 	
-	public int getSiteLibraryId() {
+	public long getSiteLibraryId() {
 		return siteLibraryId;
 	}
 
-	public void setSiteLibraryId(int siteLibraryId) {
+	public void setSiteLibraryId(long siteLibraryId) {
 		this.siteLibraryId = siteLibraryId;
 	}
